requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
numpy==1.22.2
pandas==1.4.1
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
